Add mock data and revert double data type for pricing(for now).

+++ Models/Orders/CreateOrderItemForm.cs
@@ -8,7 +8,7 @@ public class CreateOrderItemForm
     public OrderLineTypes LineType { get; set; }
     public string Details { get; set; } = string.Empty;
     public int Quantity { get; set; }
-    public decimal IndividualPrice { get; set; }
-    public decimal TotalPrice => Quantity * IndividualPrice;
+    public double IndividualPrice { get; set; }
+    public double TotalPrice => Quantity * IndividualPrice;
     
 }
